modularizes movies
