+++ server.js
@@ -3,8 +3,8 @@
 require('dotenv').config();
 const express = require('express');
 const cors = require('cors');
-const axios = require('axios');
-const handleGetWeather = require('./modules/weather'); //video said we don't need to add .js
+const handleGetWeather = require('./modules/weather');
+const handleGetMovies = require('./modules/movies');
 
 const app = express();
 
@@ -16,34 +16,4 @@ app.listen(PORT, () => console.log('server is listening on port', PORT));
 
 //routes
 app.get('/weather', handleGetWeather);
-app.get('/movie', handleGetMovie);
-
-class Movie {
-    constructor(obj){
-        this.title = obj.title;
-        this.overview = obj.overview;
-        this.average_votes = obj.vote_average;
-        this.total_votes = obj.vote_count;
-        this.image_url = `https://image.tmdb.org/t/p/w500/${obj.poster_path}`;
-        this.popularity = obj.popularity;
-        this.released_on = obj.release_date;
-    }
-}
-
-async function handleGetMovie(request, response) {
-    //Jack+Reacher format for multiple strings
-    const url = `https://api.themoviedb.org/3/search/movie?api_key=${process.env.MOVIE_API_KEY}&query=${request.query.loc}`;
-
-    try{
-        const result = await axios.get(url);
-        const movieObjArr = result.data.results.map(movieObj => new Movie(movieObj));
-        response.send(movieObjArr);
-    } catch (error) {
-        response.status(204).send({
-            message: 'No movies about this location!'
-        });
-    }
-}
-
-
-
+app.get('/movie', handleGetMovies);
